# Fix memory compression